@@ -7,9 +7,8 @@ value of claims that were filtered during the pre-processing stage.
 """
 
 from datetime import datetime
-from typing import Dict
-from google.cloud import bigquery
-from google.cloud import monitoring_v3
+
+from google.cloud import bigquery, monitoring_v3
 
 from handlers.base import Handler, HandlerBadRequestError
 from metrics import emit_gauge_metric
@@ -18,7 +17,7 @@ from metrics import emit_gauge_metric
 class PreFilteredApprovalHandler(Handler):
     """
     Calculates and emits metrics for pre-processing filter results.
-    
+
     When the pipesv2_approval pipeline completes, this handler queries BigQuery
     to aggregate claim values from the pre-processing filter stage and emits
     two metrics: one representing the total value of filtered claims, and another
@@ -26,7 +25,7 @@ class PreFilteredApprovalHandler(Handler):
     This enables monitoring of the financial impact of pre-processing filters applied
     during the approval pipeline execution.
     """
-    
+
     def __init__(
         self,
         monitoring_client: monitoring_v3.MetricServiceClient,
@@ -36,7 +35,7 @@ class PreFilteredApprovalHandler(Handler):
     ):
         """
         Initialize the handler.
-        
+
         Args:
             monitoring_client: Monitoring client (GCP or logged)
             bq_client: BigQuery client
@@ -47,17 +46,17 @@ class PreFilteredApprovalHandler(Handler):
         self.bq_client = bq_client
         self.run_project_id = run_project_id
         self.data_project_id = data_project_id
-    
-    def match(self, decoded_message: Dict) -> bool:
+
+    def match(self, decoded_message: dict) -> bool:
         """
         Determine if this handler should process the event.
-        
+
         Matches events representing successful completion of the pipesv2_approval
         pipeline, which triggers the calculation of pre-processing filter metrics.
-        
+
         Args:
             decoded_message: The decoded message dictionary to check
-            
+
         Returns:
             True if this is a pipesv2_approval pipeline completion event,
             False otherwise
@@ -65,24 +64,21 @@ class PreFilteredApprovalHandler(Handler):
         pl = decoded_message.get("payload", {})
         pipeline_uuid = pl.get("pipeline_uuid")
         pipeline_status = pl.get("status")
-        
-        return (
-            pipeline_uuid == "pipesv2_approval"
-            and pipeline_status == "COMPLETED"
-        )
-    
-    def handle(self, decoded_message: Dict) -> None:
+
+        return pipeline_uuid == "pipesv2_approval" and pipeline_status == "COMPLETED"
+
+    def handle(self, decoded_message: dict) -> None:
         """
         Calculate pre-processing filter metrics and emit to Cloud Monitoring.
-        
+
         Queries BigQuery to aggregate the total value of claims filtered during
         the pre-processing stage, then emits metrics representing both the total
         and relative values. The relative metric is calculated as the ratio of
         filtered claims value to the total value of processable claims.
-        
+
         Args:
             decoded_message: The decoded message dictionary containing pipeline completion data
-            
+
         Raises:
             HandlerBadRequestError: If the required input table variables
                 (unprocessable_claims_input_table or processable_claims_input_table)
@@ -90,35 +86,31 @@ class PreFilteredApprovalHandler(Handler):
         """
         pl = decoded_message["payload"]
         variables = pl.get("variables", {})
-        
+
         # Verify we're still handling the right event (defensive check)
         if pl.get("pipeline_uuid") != "pipesv2_approval" or pl.get("status") != "COMPLETED":
             return
-        
+
         unprocessable_table = variables.get("unprocessable_claims_input_table")
         processable_table = variables.get("processable_claims_input_table")
-        
+
         if not unprocessable_table:
             raise HandlerBadRequestError(
                 "No variable 'unprocessable_claims_input_table' found in payload."
             )
-        
+
         if not processable_table:
             raise HandlerBadRequestError(
                 "No variable 'processable_claims_input_table' found in payload."
             )
-        
+
         # Ensure fully-qualified table paths
         def ensure_full_table_ref(table: str) -> str:
-            return (
-                table
-                if "." in table
-                else f"{self.data_project_id}.{table}"
-            )
-        
+            return table if "." in table else f"{self.data_project_id}.{table}"
+
         full_unprocessable_table = ensure_full_table_ref(unprocessable_table)
         full_processable_table = ensure_full_table_ref(processable_table)
-        
+
         # Single BigQuery query to get both values
         query = f"""
         WITH unprocessable_sum AS (
@@ -129,21 +121,21 @@ class PreFilteredApprovalHandler(Handler):
             SELECT COALESCE(SUM(vl_pago), 0) AS sum_processable_vl_pago
             FROM `{full_processable_table}`
         )
-        SELECT 
+        SELECT
             u.total_vl_pago,
             p.sum_processable_vl_pago,
-            CASE 
-                WHEN p.sum_processable_vl_pago > 0 
-                THEN u.total_vl_pago / p.sum_processable_vl_pago 
-                ELSE 0 
+            CASE
+                WHEN p.sum_processable_vl_pago > 0
+                THEN u.total_vl_pago / p.sum_processable_vl_pago
+                ELSE 0
             END AS relative_vl_pago
         FROM unprocessable_sum u, processable_sum p
         """
-        
+
         query_job = self.bq_client.query(query)
         result = query_job.result()
         row = next(result, None)
-        
+
         total_vl_pago = float(row.total_vl_pago or 0.0)
         relative_vl_pago = float(row.relative_vl_pago or 0.0)
 
@@ -151,20 +143,18 @@ class PreFilteredApprovalHandler(Handler):
         labels = {
             "approved": "true",
         }
-        
+
         # Partner label is required
         partner_value = variables.get("partner")
         if not partner_value:
-            raise HandlerBadRequestError(
-                "Missing required 'partner' variable in payload."
-            )
+            raise HandlerBadRequestError("Missing required 'partner' variable in payload.")
         labels["partner"] = str(partner_value)
-        
+
         # Parse timestamp
         source_timestamp = datetime.fromisoformat(
             decoded_message["source_timestamp"].replace("Z", "+00:00")
         )
-        
+
         # Emit metric representing total value of claims filtered in pre-processing
         emit_gauge_metric(
             monitoring_client=self.monitoring_client,
@@ -174,7 +164,7 @@ class PreFilteredApprovalHandler(Handler):
             labels=labels,
             timestamp=source_timestamp,
         )
-        
+
         # Emit metric representing relative value of claims filtered in pre-processing
         emit_gauge_metric(
             monitoring_client=self.monitoring_client,
@@ -184,4 +174,3 @@ class PreFilteredApprovalHandler(Handler):
             labels=labels,
             timestamp=source_timestamp,
         )
-
