@@ -3,16 +3,16 @@
 import logging
 
 import functions_framework
-from flask import make_response
 from cloudevents.http import CloudEvent
+from flask import make_response
 
+from bigquery import create_bigquery_client
+from config import Config
 from dispatcher import HandlerDispatcher
 from handlers.pipeline_run import PipelineRunHandler
 from handlers.pre_filtered_approval import PreFilteredApprovalHandler
 from handlers.pre_filtered_wrangling import PreFilteredWranglingHandler
-from metrics import emit_gauge_metric, create_monitoring_client
-from bigquery import create_bigquery_client
-from config import Config
+from metrics import create_monitoring_client
 
 logging.basicConfig(level=logging.INFO)
 
@@ -24,12 +24,12 @@ def create_handlers(
 ) -> list:
     """
     Create handler instances with injected dependencies.
-    
+
     Args:
         monitoring_client: Monitoring client (GCP or logged)
         bq_client: BigQuery client
         config: Application configuration
-        
+
     Returns:
         List of handler instances
     """
@@ -58,20 +58,20 @@ _dispatcher = None
 
 
 @functions_framework.cloud_event
-def log_and_metric_pubsub(cloud_event: CloudEvent):
+def handle_cloud_event(cloud_event: CloudEvent):
     """
     Triggered by Pub/Sub message (production) or HTTP request (local).
-    
+
     The same handler works for both modes:
     - Production: Receives CloudEvent from Pub/Sub subscription
     - Local: functions-framework automatically converts HTTP POST requests to CloudEvent format
-    
+
     Routes to appropriate handlers based on event content.
-    
+
     Uses lazy initialization to avoid macOS fork safety issues with Google Cloud clients.
     """
     global _dispatcher
-    
+
     # Initialize on first request (happens after fork in each worker)
     if _dispatcher is None:
         config = Config()
@@ -79,7 +79,7 @@ def log_and_metric_pubsub(cloud_event: CloudEvent):
         bq_client = create_bigquery_client(config)
         handlers = create_handlers(monitoring_client, bq_client, config)
         _dispatcher = HandlerDispatcher(handlers)
-    
+
     try:
         return _dispatcher.dispatch(cloud_event)
 
