@@ -9,8 +9,9 @@ import logging
 import os
 import sys
 import time
-from google.cloud import bigquery
+
 from google.api_core import client_options
+from google.cloud import bigquery
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -20,11 +21,9 @@ def create_bigquery_client(project_id: str) -> bigquery.Client:
     """Create a BigQuery client configured for local emulator."""
     emulator_host = os.getenv("BIGQUERY_EMULATOR_HOST", "localhost:9050")
     emulator_endpoint = f"http://{emulator_host}"
-    
-    client_options_obj = client_options.ClientOptions(
-        api_endpoint=emulator_endpoint
-    )
-    
+
+    client_options_obj = client_options.ClientOptions(api_endpoint=emulator_endpoint)
+
     return bigquery.Client(
         project=project_id,
         client_options=client_options_obj,
@@ -53,12 +52,12 @@ def wait_for_emulator(client: bigquery.Client, max_retries: int = 30, delay: flo
 
 def seed_bigquery_data(client: bigquery.Client, project_id: str) -> None:
     """Create datasets and tables with sample data for local testing."""
-    
+
     # Create tmp_abertta dataset
     dataset_id = "tmp_abertta"
     dataset = bigquery.Dataset(f"{project_id}.{dataset_id}")
     dataset.location = "US"
-    
+
     try:
         logger.info(f"Creating dataset: {dataset_id}")
         client.create_dataset(dataset, exists_ok=True)
@@ -66,24 +65,24 @@ def seed_bigquery_data(client: bigquery.Client, project_id: str) -> None:
     except Exception as e:
         logger.error(f"Failed to create dataset {dataset_id}: {e}")
         raise
-    
+
     # Define table schema for claims tables
     schema = [
         bigquery.SchemaField("vl_pago", "FLOAT", mode="NULLABLE"),
     ]
-    
+
     # Table names from README example
     unprocessable_table_name = "unprocessable_claims_abertta_20251105_171427"
     processable_table_name = "processable_claims_abertta_20251105_171427"
-    
+
     # Create unprocessable claims table
     unprocessable_table_ref = f"{project_id}.{dataset_id}.{unprocessable_table_name}"
     unprocessable_table = bigquery.Table(unprocessable_table_ref, schema=schema)
-    
+
     try:
         logger.info(f"Creating table: {unprocessable_table_name}")
         client.create_table(unprocessable_table, exists_ok=True)
-        
+
         # Insert sample data: total vl_pago = 1500.0
         unprocessable_rows = [
             {"vl_pago": 500.0},
@@ -98,15 +97,15 @@ def seed_bigquery_data(client: bigquery.Client, project_id: str) -> None:
     except Exception as e:
         logger.error(f"Failed to create/seed table {unprocessable_table_name}: {e}")
         raise
-    
+
     # Create processable claims table
     processable_table_ref = f"{project_id}.{dataset_id}.{processable_table_name}"
     processable_table = bigquery.Table(processable_table_ref, schema=schema)
-    
+
     try:
         logger.info(f"Creating table: {processable_table_name}")
         client.create_table(processable_table, exists_ok=True)
-        
+
         # Insert sample data: total vl_pago = 8000.0
         processable_rows = [
             {"vl_pago": 2000.0},
@@ -121,7 +120,7 @@ def seed_bigquery_data(client: bigquery.Client, project_id: str) -> None:
     except Exception as e:
         logger.error(f"Failed to create/seed table {processable_table_name}: {e}")
         raise
-    
+
     logger.info("BigQuery seeding completed successfully")
     logger.info(f"Dataset: {project_id}.{dataset_id}")
     logger.info(f"Tables: {unprocessable_table_name}, {processable_table_name}")
@@ -131,16 +130,16 @@ def main():
     """Main entry point for seeding script."""
     project_id = os.getenv("BIGQUERY_PROJECT_ID", "test-project")
     emulator_host = os.getenv("BIGQUERY_EMULATOR_HOST", "localhost:9050")
-    
+
     logger.info(f"Seeding BigQuery emulator at {emulator_host}")
     logger.info(f"Using project: {project_id}")
-    
+
     client = create_bigquery_client(project_id)
-    
+
     if not wait_for_emulator(client):
         logger.error("BigQuery emulator is not ready. Exiting.")
         sys.exit(1)
-    
+
     try:
         seed_bigquery_data(client, project_id)
         logger.info("Seeding completed successfully")
@@ -151,4 +150,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
